% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tile-methods.R
\name{approxGeneticMap}
\alias{approxGeneticMap}
\title{This uses a monotonic polynomial regression (9th order by default) to smooth
over the genentic map markers}
\usage{
approxGeneticMap(genmap, degree = 9)
}
\description{
This uses a monotonic polynomial regression (9th order by default) to smooth
over the genentic map markers
}

