% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/parentage-methods.R
\name{caclulateLODs}
\alias{caclulateLODs}
\title{calculate LODs for QQ/UU and QU/UU, for maximum parent 1 and 2}
\usage{
caclulateLODs(x)
}
\description{
calculate LODs for QQ/UU and QU/UU, for maximum parent 1 and 2
}

