% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utilities.R
\docType{methods}
\name{calcFreqs,ProgenyArray-method}
\alias{calcFreqs,ProgenyArray-method}
\title{Calculate allele frequencies from parents in a ProgenyArray}
\usage{
\S4method{calcFreqs}{ProgenyArray}(x)
}
\arguments{
\item{x}{a ProgenyArray object}
}
\description{
Calculate allele frequencies from parents in a ProgenyArray
}

