% Generated by roxygen2 (4.0.1): do not edit by hand
\name{checkValidProgenyArray}
\alias{checkValidProgenyArray}
\title{Check if is valid: all progeny have mothers in genotype matrix.}
\usage{
checkValidProgenyArray(x)
}
\description{
Check if is valid: all progeny have mothers in genotype matrix.
}

