% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/phasing.R
\name{emHaplotypeImpute}
\alias{emHaplotypeImpute}
\title{Use EM Haplotype Phasing by Imputation Algorithm (internal function)}
\usage{
emHaplotypeImpute(pgeno, fgeno, other_parents, freqs, ehet, ehom,
  free_pi = FALSE, na_thresh = 0.8, init = NULL, extra = FALSE,
  max_iter = 60L, eps = 1e-09)
}
\arguments{
\item{pgeno}{progeny genotypes}

\item{fgeno}{father (or other parent) genotypes}

\item{other_parents}{indices to corresponding father genotype columns in fgeno}

\item{freqs}{allele frequencies}

\item{ehet}{heterozygous error rate}

\item{ehom}{homozygous error rate}

\item{free_pi}{whether to allow pi to vary (can cause identifiability issues}

\item{na_thresh}{how much missingness to tolerate before pruning individual}

\item{init}{optional initial reponsibilities (advanced use)}

\item{extra}{include optional debugging output (advanced use)}

\item{max_iter}{maximum number of EM iterations before quiting}

\item{eps}{epsilon in log likelihood before calling convergence}
}
\description{
Use EM Haplotype Phasing by Imputation Algorithm (internal function)
}

