% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/phasing-utilities.R
\name{emPlot}
\alias{emPlot}
\title{A diagnostic plot of haplotype imputation of simulated data}
\usage{
emPlot(res, sim)
}
\arguments{
\item{res}{phasing results}

\item{sim}{simulation data}
}
\description{
A diagnostic plot of haplotype imputation of simulated data
}

