% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/phasing.R
\name{extractProgenyHaplotypes}
\alias{extractProgenyHaplotypes}
\title{Extract all progeny haplotypes from all parents}
\usage{
extractProgenyHaplotypes(x, included_parents)
}
\description{
Extract all progeny haplotypes from all parents
}

