% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/progenyarray-methods.R
\docType{methods}
\name{fathers,ProgenyArray-method}
\alias{fathers,ProgenyArray-method}
\title{Accessor for fathers in a ProgenyArray object}
\usage{
\S4method{fathers}{ProgenyArray}(x)
}
\arguments{
\item{x}{a ProgenyArray object}
}
\description{
Internally, this references the \code{parents} slot.
}

