% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/phasing.R
\name{filterSibFamilies}
\alias{filterSibFamilies}
\title{Filter sibling families by minimum number of offspring}
\usage{
filterSibFamilies(sibfamilies, min_child)
}
\description{
Filter sibling families by minimum number of offspring
}

