% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utilities.R
\docType{methods}
\name{freqs,ProgenyArray-method}
\alias{freqs,ProgenyArray-method}
\title{Getter for allele frequencies}
\usage{
\S4method{freqs}{ProgenyArray}(x)
}
\arguments{
\item{x}{a ProgenyArray object}
}
\description{
Getter for allele frequencies
}

