% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tile-methods.R
\name{geneticTiles}
\alias{geneticTiles}
\title{Create a PhasingTiles object for tiles based on genetic distances}
\usage{
geneticTiles(x, mapfile, length)
}
\arguments{
\item{x}{a ProgenyArray object}

\item{mapfile}{a tab-delimited file (no header) of marker, chromosome, position, cumulative genetic distance}

\item{length}{length of each tile in centiMorgans (actual windows may be a bit larger than this)}
}
\description{
Create a PhasingTiles object for tiles based on genetic distances
}

