% Generated by roxygen2 (4.0.1): do not edit by hand
\name{genotypeFreqs}
\alias{genotypeFreqs}
\title{Calculate genotype frequencies.}
\usage{
genotypeFreqs(x, counts = FALSE)
}
\arguments{
\item{x}{a genotype matrix}

\item{counts}{whether to return genotype counts rather than frequency}
}
\description{
Calculate genotype frequencies.
}

