% Generated by roxygen2 (4.0.1): do not edit by hand
\name{genotypingErrorMatrix}
\alias{genotypingErrorMatrix}
\title{Create a genotyping error matrix.}
\usage{
genotypingErrorMatrix(ehet = 0.6, ehom = 0.1)
}
\arguments{
\item{ehet}{the probability that a heterozygous genotype is incorrect}

\item{ehom}{the probability that a homozygous genotype is incorrect

This function creates a matrix of the form:
  [1-e, e/2, e/2,
   E/2, 1-E, E/2,
   e/2, e/2, 1-e]
where E is the probability the heterozygous genotype call is incorrect, and
e is the probability that the homozygous genotypes call is incorrect.}
}
\description{
Returns a matrix of transition probabilities for a given genotype given error.
}

