% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/progenyarray-methods.R
\docType{methods}
\name{granges,ProgenyArray-method}
\alias{granges,ProgenyArray-method}
\title{Accessor for genomic ranges from a ProgenyArray object}
\usage{
\S4method{granges}{ProgenyArray}(x, use.mcols = FALSE, ...)
}
\arguments{
\item{x}{a ProgenyArray object}
}
\description{
Accessor for genomic ranges from a ProgenyArray object
}

