% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/phasing-utilities.R
\name{hapAgreementStats}
\alias{hapAgreementStats}
\title{Agreement Rate (across both haplotypes) for all windows}
\usage{
hapAgreementStats(ph1, ph2, x = NULL, tiles = NULL, chrom = NULL)
}
\arguments{
\item{ph1}{phasing results from haplotype imputation 1}

\item{ph2}{phasing results from haplotype imputation 2}
}
\description{
Agreement Rate (across both haplotypes) for all windows
}

