% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{inferFathers,ProgenyArray-method}
\alias{inferFathers,ProgenyArray-method}
\title{Infer Fathers for all progeny when the mother is known}
\usage{
\S4method{inferFathers}{ProgenyArray}(x, ehet, ehom, verbose = FALSE)
}
\description{
Infer Fathers for all progeny when the mother is known
}

