% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{inferParents,SimulatedProgenyArray-method}
\alias{inferParents,SimulatedProgenyArray-method}
\title{Infer parents for SimulatedProgenyArray}
\usage{
\S4method{inferParents}{SimulatedProgenyArray}(x, ehet, ehom, verbose = FALSE)
}
\arguments{
\item{x}{a SimulatedProgenyArray object}

\item{ehet}{heterozygous error rate}

\item{ehom}{homozygous error rate}
}
\description{
Infer parents for SimulatedProgenyArray
}

