% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{isSelfed,ProgenyArray-method}
\alias{isSelfed,ProgenyArray-method}
\title{Method to check if individual is selfed}
\usage{
\S4method{isSelfed}{ProgenyArray}(x)
}
\arguments{
\item{x}{a ProgenyArray object}
}
\description{
Method to check if individual is selfed
}

