% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tile-methods.R
\name{loadGeneticMap}
\alias{loadGeneticMap}
\title{Load genetic map of four columns: marker, chrom, position, interval, cumulative}
\usage{
loadGeneticMap(mapfile)
}
\description{
Load genetic map of four columns: marker, chrom, position, interval, cumulative
}

