% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utilities.R
\name{miMatrixPerTile}
\alias{miMatrixPerTile}
\title{Trio Mendelian inconsistency across full-sib family per phased tile}
\usage{
miMatrixPerTile(x, parent_1, parent_2, chrom, tiles, phased_1 = NULL,
  phased_2 = NULL)
}
\arguments{
\item{parent_1}{parent 1}

\item{parent_2}{parent 2}

\item{chrom}{which chromosome was used}

\item{tiles}{tiles used in phasing}

\item{phased_1}{phasing results for parent 1 (if \code{NULL}, uses genotype data)}

\item{phased_2}{phasing results for parent 2 (if \code{NULL}, uses genotype data)}
}
\description{
Trio Mendelian inconsistency across full-sib family per phased tile
}

