% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utilities.R
\name{miRates}
\alias{miRates}
\title{Mendelian inconsistency rates across a chromosome}
\usage{
miRates(mi, x, chrom, tiles)
}
\arguments{
\item{mi}{results from \code{miMatrixPerTile}}

\item{x}{ProgenyArray object}

\item{chrom}{which chromosome was used}

\item{tiles}{tiles used in phasing}
}
\description{
Mendelian inconsistency rates across a chromosome
}

