% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/progenyarray-methods.R
\docType{methods}
\name{mothers,ProgenyArray-method}
\alias{mothers,ProgenyArray-method}
\title{Accessor for mothers in a ProgenyArray object}
\usage{
\S4method{mothers}{ProgenyArray}(x)
}
\arguments{
\item{x}{a ProgenyArray object}
}
\description{
Internally, this references the \code{parents} slot.
}

