% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/progenyarray-methods.R
\docType{methods}
\name{parentGenotypes,ProgenyArray-method}
\alias{parentGenotypes,ProgenyArray-method}
\title{Accessor for parent genotypes}
\usage{
\S4method{parentGenotypes}{ProgenyArray}(x, seqname = NULL)
}
\arguments{
\item{x}{a ProgenyArray object}
}
\description{
Accessor for parent genotypes
}

