% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/simulation-methods.R
\docType{methods}
\name{parentGenotypes,SimulatedProgenyArray-method}
\alias{parentGenotypes,SimulatedProgenyArray-method}
\title{Accessor for parent genotypes}
\usage{
\S4method{parentGenotypes}{SimulatedProgenyArray}(x)
}
\arguments{
\item{x}{a SimulationData object}
}
\description{
Accessor for parent genotypes
}

