% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/progenyarray-methods.R
\docType{methods}
\name{parentLLPlot,ProgenyArray-method}
\alias{parentLLPlot,ProgenyArray-method}
\title{Create a diagnostic plot of a parentage inference, looking at all log likelihood ratios.}
\usage{
\S4method{parentLLPlot}{ProgenyArray}(x, progeny)
}
\arguments{
\item{x}{a ProgenyArray object}

\item{progeny}{which progeny to use (an index to its column in progeny genotypes)}
}
\description{
Create a diagnostic plot of a parentage inference, looking at all log likelihood ratios.
}

