% Generated by roxygen2 (4.0.1): do not edit by hand
\name{parentLLPlot}
\alias{parentLLPlot}
\title{Create a diagnostic plot of a parentage inference, looking at all log likelihood ratios.}
\usage{
parentLLPlot(x, progeny)
}
\description{
Create a diagnostic plot of a parentage inference, looking at all log likelihood ratios.
}

