% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/progenyarray-methods.R
\docType{methods}
\name{parentNames,ProgenyArray-method}
\alias{parentNames,ProgenyArray-method}
\title{Return parent sample names}
\usage{
\S4method{parentNames}{ProgenyArray}(x)
}
\arguments{
\item{x}{a ProgenyArray object}
}
\description{
Return parent sample names
}

