% Generated by roxygen2 (4.0.1): do not edit by hand
\name{parentNames}
\alias{parentNames}
\title{Accessor for parents sample names}
\usage{
parentNames(x)
}
\arguments{
\item{x}{a ProgenyArray object}

\item{value}{vector of parent names}
}
\description{
Accessor for parents sample names

Set method for parents sample names
}

