% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/progenyarray-methods.R
\docType{methods}
\name{parents,ProgenyArray-method}
\alias{parents,ProgenyArray-method}
\title{Accessor for parents \code{data.frame} in a ProgenyArray object}
\usage{
\S4method{parents}{ProgenyArray}(x, use_names = FALSE)
}
\arguments{
\item{x}{a ProgenyArray object}
}
\description{
Accessor for parents \code{data.frame} in a ProgenyArray object
}

