% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/phasing.R
\docType{methods}
\name{phaseParents,ProgenyArray-method}
\alias{phaseParents,ProgenyArray-method}
\title{Phase all Parents in a ProgenyArray object}
\usage{
\S4method{phaseParents}{ProgenyArray}(x, tiles, ehet = 0.8, ehom = 0.1,
  na_thresh = 0.8, min_child = 8, verbose = TRUE)
}
\arguments{
\item{x}{a ProgenyArray object}

\item{tiles}{a PhasingTiles object}

\item{ehet}{heterozygous error rate}

\item{ehom}{homozygous error rate}

\item{na_thresh}{how much missingness to tolerate before pruning individual}

\item{min_child}{minimum number of children to include a parent for phasing}

\item{verbose}{report status messages}
}
\description{
Phase all Parents in a ProgenyArray object
}

