% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{phaseParents}
\alias{phaseParents}
\title{Phase all parents data in a ProgenyArray}
\usage{
phaseParents(x, tiles, ehet = 0.8, ehom = 0.1, na_thresh = 0.8,
  min_child = 8, verbose = TRUE)
}
\description{
Phase all parents data in a ProgenyArray
}

