% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/phasing.R
\docType{methods}
\name{phases,ProgenyArray-method}
\alias{phases,ProgenyArray-method}
\title{Merge phasing data}
\usage{
\S4method{phases}{ProgenyArray}(x, include_ll = FALSE, verbose = TRUE)
}
\arguments{
\item{x}{a ProgenyArray object with phased parents}

\item{include_ll}{a logical indicating whether to include the haplotype log-likelihoods}
}
\description{
Merge phasing data
}

