% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{phases}
\alias{phases}
\title{Output all a simplified table of phasing data}
\usage{
phases(x, include_ll = FALSE, verbose = TRUE)
}
\description{
Output all a simplified table of phasing data
}

