% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/phasing.R
\name{phaseSibFamily}
\alias{phaseSibFamily}
\title{Phase Sibling Family by Haplotype Imputation}
\usage{
phaseSibFamily(x, sibfam, chrom, tiles, ehet = 0.8, ehom = 0.1,
  na_thresh = 0.8)
}
\arguments{
\item{x}{a ProgenyArray object}

\item{sibfam}{a dataframe from \code{sibFamily()}}

\item{chrom}{which chromosome to use}

\item{tiles}{a PhasingTiles object}

\item{ehet}{heterozygous error rate}

\item{ehom}{homozygous error rate}

\item{na_thresh}{missingness threshold to remove individual from clustering}
}
\description{
Phase Sibling Family by Haplotype Imputation
}

