% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{PhasingTiles-class}
\alias{PhasingTiles-class}
\title{A container for tiles used for phasing by imputation}
\description{
A container for tiles used for phasing by imputation
}
\section{Slots}{

\describe{
\item{\code{type}}{a character describing the type of tile (position, SNP, genetic)}

\item{\code{width}}{the width of each tile}

\item{\code{tiles}}{a list of lists, with the first list containing all chromosomes' tiles, and the second list containing the indices of the SNPs per tile}

\item{\code{info}}{a list containing supplementary info about tiles, e.g genetic map and related smoothed versions}
}}

