% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tile-methods.R
\name{physicalTiles}
\alias{physicalTiles}
\title{Create a PhasingTiles object for tiles based on physical distances}
\usage{
physicalTiles(x, width)
}
\arguments{
\item{x}{a ProgenyArray object}

\item{width}{width in bases of each tile (except last)}
}
\description{
Create a PhasingTiles object for tiles based on physical distances
}

