% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tile-methods.R
\name{plotGeneticMap}
\alias{plotGeneticMap}
\title{Function to inspect the Montonic Polynomial Regression Smoothing of Genetic Map}
\usage{
plotGeneticMap(tiles)
}
\arguments{
\item{tiles}{a PhasingTiles object}
}
\description{
Function to inspect the Montonic Polynomial Regression Smoothing of Genetic Map
}

