% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{possibleFathers,ProgenyArray-method}
\alias{possibleFathers,ProgenyArray-method}
\title{Accessor for possible fathers in a ProgenyArray object}
\usage{
\S4method{possibleFathers}{ProgenyArray}(x)
}
\arguments{
\item{x}{a ProgenyArray object}
}
\description{
Accessor for possible fathers in a ProgenyArray object
}

