% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tile-methods.R
\name{predictGeneticMap}
\alias{predictGeneticMap}
\title{Use Monotonic Polynomial Regression Fits to Predict Genetic Distance Between
Arbitary Markers}
\usage{
predictGeneticMap(approx, ranges, as_df = TRUE)
}
\description{
Use Monotonic Polynomial Regression Fits to Predict Genetic Distance Between
Arbitary Markers
}

