% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{ProgenyArray-class}
\alias{ProgenyArray-class}
\title{An S4 class that stores genotyping data from a progeny array design.}
\description{
An S4 class that stores genotyping data from a progeny array design.
}
\section{Slots}{

\describe{
\item{\code{ranges}}{a \code{GenomicRanges} object of loci}

\item{\code{ref}}{reference alleles}

\item{\code{alt}}{alternate alleles}

\item{\code{progeny_geno}}{a matrix of bialleic progeny genotypes}

\item{\code{parents_geno}}{a matrix of bialleic parents genotypes}

\item{\code{supplied_mothers}}{integer vector indicating the mother of each progeny, supplied by user}

\item{\code{mothers}}{integer vector indicating the inferred mother of each progeny}

\item{\code{fathers}}{integer vector indicating the inferred father of each progeny}

\item{\code{parents}}{dataframe with inferred parents and model information}

\item{\code{complete_loci}}{which loci are not all missing or fixed in the parents}

\item{\code{parent_lods}}{matrices from C++ parentage routine (for debugging)}

\item{\code{lodcutoff}}{the LOD cutoff used in parentage inference}

\item{\code{phased_parents}}{a list of parents' phased tiles for all chromosomes}

\item{\code{phased_parents_metadata}}{a list of phasing parameters}

\item{\code{tiles}}{a list of tiles for all chromosomes}

\item{\code{sibfams}}{a list of sibling families for all phased parents}
}}

