% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ProgenyArray-package.r, R/progenyarray-methods.R
\docType{package}
\name{ProgenyArray}
\alias{ProgenyArray}
\alias{ProgenyArray-package}
\title{ProgenyArray.}
\usage{
ProgenyArray(progeny_geno, parents_geno, mothers = integer(),
  loci = GRanges(), ref = character(), alt = character())
}
\arguments{
\item{progeny_geno}{progeny genotypes matrix (all entries should be 0, 1, 2, or NA)}

\item{parents_geno}{parents genotypes matrix (all entries should be 0, 1, 2, or NA)}

\item{mothers}{user-supplied mothers (e.g. known from collection)}

\item{loci}{a \code{GRanges} object of loci}

\item{ref}{a character vector of reference loci}

\item{alt}{a character vector of alternate loci}
}
\description{
ProgenyArray.

Create a new ProgenyArray object for storing genotyping and other data from
a progeny array experiment.
}

