% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/progenyarray-methods.R
\docType{methods}
\name{progenyGenotypes,ProgenyArray-method}
\alias{progenyGenotypes,ProgenyArray-method}
\title{Accessor for progeny genotypes}
\usage{
\S4method{progenyGenotypes}{ProgenyArray}(x, seqname = NULL)
}
\arguments{
\item{x}{a ProgenyArray object}
}
\description{
Accessor for progeny genotypes
}

