% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{progenyGenotypes,SimulatedProgenyArray-method}
\alias{progenyGenotypes,SimulatedProgenyArray-method}
\title{Accessor for progeny genotypes}
\usage{
\S4method{progenyGenotypes}{SimulatedProgenyArray}(x)
}
\arguments{
\item{x}{a SimulationData object}
}
\description{
Accessor for progeny genotypes
}

