% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/progenyarray-methods.R
\docType{methods}
\name{progenyNames,ProgenyArray-method}
\alias{progenyNames,ProgenyArray-method}
\title{Return progeny sample names}
\usage{
\S4method{progenyNames}{ProgenyArray}(x)
}
\arguments{
\item{x}{a ProgenyArray object}
}
\description{
Return progeny sample names
}

