% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{progenyNames}
\alias{progenyNames}
\title{Accessor for progeny sample names}
\usage{
progenyNames(x)

\S4method{parentage}{ProgenyArray}(x)
}
\arguments{
\item{x}{a ProgenyArray object}

\item{value}{vector of parent names}

\item{x}{a ProgenyArray object}
}
\description{
Accessor for progeny sample names

Set method for progeny sample names

Get inferred parents as a dataframe
}

