% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{ref,ProgenyArray-method}
\alias{ref,ProgenyArray-method}
\title{Accessor for reference alleles from a ProgenyArray object}
\usage{
\S4method{ref}{ProgenyArray}(x)
}
\arguments{
\item{x}{a ProgenyArray object}
}
\description{
Accessor for reference alleles from a ProgenyArray object
}

