% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{samples,ProgenyArray-method}
\alias{samples,ProgenyArray-method}
\title{Accessor for samples from a ProgenyArray object}
\usage{
\S4method{samples}{ProgenyArray}(object)
}
\arguments{
\item{object}{a ProgenyArray object}
}
\description{
Accessor for samples from a ProgenyArray object
}

