% Generated by roxygen2 (4.0.1): do not edit by hand
\name{set-methods}
\alias{set-methods}
\title{Set method for fathers}
\arguments{
\item{object}{}

\item{value}{}

\item{object}{}

\item{value}{}

\item{object}{}

\item{value}{}
}
\description{
Set method for fathers

Set method for mothers

Set method for progeny
}

