% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tile-methods.R
\docType{methods}
\name{show,PhasingTiles-method}
\alias{show,PhasingTiles-method}
\title{Pretty-print a PhasingTiles object}
\usage{
\S4method{show}{PhasingTiles}(object)
}
\arguments{
\item{object}{a PhasingTiles object}
}
\description{
Pretty-print a PhasingTiles object
}

