% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{show,ProgenyArray-method}
\alias{show,ProgenyArray-method}
\title{Pretty-print a ProgenyArray object}
\usage{
\S4method{show}{ProgenyArray}(object)
}
\arguments{
\item{object}{a ProgenyArray object}
}
\description{
Pretty-print a ProgenyArray object
}

