% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/haplotype-simulation.R, R/phasing.R
\name{sibFamily}
\alias{sibFamily}
\title{Simulate a half-sib family, fixed number of sites}
\usage{
sibFamily(x, parent)

sibFamily(x, parent)
}
\arguments{
\item{x}{a ProgenyArray object}

\item{parent}{the parent to bring the half sib family}

\item{n}{size of halfsib family}

\item{nsites}{number of sites}

\item{rates}{selfing, halfsib, fullsib proportions}

\item{nfullsib_fathers}{number of fathers contributing to fullsib families}

\item{father_F}{proportion of fathers that are inbred}

\item{mom_inbred}{whether mom is inbred}

\item{ehet}{heterozygous error rate}

\item{ehom}{homozygous error rate}
}
\value{
a list full of simulation data
}
\description{
Simulate a half-sib family, fixed number of sites

Create a sibling family for a single parent.
}

