% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{SimulatedProgenyArray-class}
\alias{SimulatedProgenyArray-class}
\title{An S4 class that wraps a ProgenyArray object of noisy, missing data}
\description{
An S4 class that wraps a ProgenyArray object of noisy, missing data
}
\section{Slots}{

\describe{
\item{\code{nparent}}{parent population size}

\item{\code{nprogeny}}{progeny population size}

\item{\code{nloci}}{number of loci}

\item{\code{selfing}}{selfing rate}

\item{\code{prop_parent_missing}}{proportion of genotype data that are missing in parents}

\item{\code{prop_progeny_missing}}{proportion of genotype data that are missing in progeny}

\item{\code{ehet}}{heterozygous error rate}

\item{\code{ehom}}{homozygous error rate}

\item{\code{progeny_geno}}{full genotype matrix for progeny}

\item{\code{parent_geno}}{full genotype matrix for parents}

\item{\code{mothers}}{true mothers of each progeny}

\item{\code{fathers}}{true mothers of each progeny}

\item{\code{progeny_array}}{ProgenyArray object of these data, with missing values and errors}
}}

