% Generated by roxygen2 (4.0.1): do not edit by hand
\name{SimulatedProgenyArray}
\alias{SimulatedProgenyArray}
\title{Create some simulation data}
\usage{
SimulatedProgenyArray(nparent, nprogeny, nloci, selfing = 0.5,
  prop_parent_missing = 0.2, prop_progeny_missing = 0.2, ehet = 0.8,
  ehom = 0.1)
}
\arguments{
\item{nparent}{}

\item{nprogeny}{}

\item{nloci}{}

\item{prop_parent_missing}{}

\item{prop_progeny_missing}{}

\item{ehet}{}

\item{ehom}{}
}
\description{
Create some simulation data
}

