% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tile-methods.R
\name{snpTiles}
\alias{snpTiles}
\title{Create a PhasingTiles object for tiles based on fixed number of SNPs per tile}
\usage{
snpTiles(x, nsnps)
}
\arguments{
\item{x}{a ProgenyArray object}

\item{nsnps}{number of SNPs per tile}
}
\description{
Create a PhasingTiles object for tiles based on fixed number of SNPs per tile
}

