% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/progenyarray-methods.R
\docType{methods}
\name{suppliedMothers,ProgenyArray-method}
\alias{suppliedMothers,ProgenyArray-method}
\title{Accessor for user-supplied mothers}
\usage{
\S4method{suppliedMothers}{ProgenyArray}(x)
}
\arguments{
\item{x}{a ProgenyArray object}
}
\description{
Accessor for user-supplied mothers
}

