% Generated by roxygen2 (4.0.1): do not edit by hand
\name{suppliedMothers}
\alias{suppliedMothers}
\title{Accessor for user-supplied mothers (returning indices or names)}
\usage{
suppliedMothers(x)
}
\arguments{
\item{x}{a ProgenyArray object}

\item{value}{vector of mother indices}
}
\description{
Accessor for user-supplied mothers (returning indices or names)

Set method for supplied mothers
}

