% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/phasing.R
\name{tileRef}
\alias{tileRef}
\title{A function that gets the reference alleles used in the tiles.  This is an
important difference from ref() which returns all reference alleles.}
\usage{
tileRef(x)
}
\description{
A function that gets the reference alleles used in the tiles.  This is an
important difference from ref() which returns all reference alleles.
}

