% Generated by roxygen2 (4.0.1): do not edit by hand
\name{transmissionMatrix}
\alias{transmissionMatrix}
\title{Create a Mendelian transmission matrix, with error in observed progeny genotypes.}
\usage{
transmissionMatrix(ehet = 0.6, ehom = 0.1)
}
\arguments{
\item{ehet}{the probability that a heterozygous genotype is incorrect}

\item{ehom}{the probability that a homozygous genotype is incorrect}
}
\description{
Create a Mendelian transmission matrix, with error in observed progeny genotypes.
}

