% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utilities.R
\name{alleleFreqs}
\alias{alleleFreqs}
\title{Calculate allele frequenceies.}
\usage{
alleleFreqs(x, min = FALSE)
}
\arguments{
\item{x}{a genotype matrix}

\item{min}{force minimum allele frequency to be 1/nsample}
}
\description{
Calculate allele frequenceies.
}

