% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{alt,ProgenyArray-method}
\alias{alt,ProgenyArray-method}
\title{Accessor for alterate alleles from a ProgenyArray object}
\usage{
\S4method{alt}{ProgenyArray}(x)
}
\arguments{
\item{x}{a ProgenyArray object}
}
\description{
Accessor for alterate alleles from a ProgenyArray object
}

