% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{fathers,ProgenyArray-method}
\alias{fathers,ProgenyArray-method}
\title{Accessor for fathers in a ProgenyArray object}
\usage{
\S4method{fathers}{ProgenyArray}(x)
}
\arguments{
\item{x}{a ProgenyArray object}
}
\description{
Accessor for fathers in a ProgenyArray object
}

