% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{granges,ProgenyArray-method}
\alias{granges,ProgenyArray-method}
\title{Accessor for genomic ranges from a ProgenyArray object}
\usage{
\S4method{granges}{ProgenyArray}(x, use.mcols = FALSE, ...)
}
\arguments{
\item{x}{a ProgenyArray object}

\item{use.mcols}{
    \code{TRUE} or \code{FALSE} (the default).
    Whether the metadata columns on \code{x} (accessible with \code{mcols(x)})
    should be propagated to the returned object or not.
  }

\item{...}{
    Additional arguments, for use in specific methods.
  }
}
\description{
Accessor for genomic ranges from a ProgenyArray object
}

