% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/progenyarray-methods.R
\docType{methods}
\name{isSelfed,ProgenyArray-method}
\alias{isSelfed,ProgenyArray-method}
\title{Method to check if individual is selfed}
\usage{
\S4method{isSelfed}{ProgenyArray}(x)
}
\arguments{
\item{x}{a ProgenyArray object}
}
\description{
Method to check if individual is selfed
}

