% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{parentGenotypes,ProgenyArray-method}
\alias{parentGenotypes,ProgenyArray-method}
\title{Accessor for parent genotypes}
\usage{
\S4method{parentGenotypes}{ProgenyArray}(x)
}
\arguments{
\item{x}{a ProgenyArray object}
}
\description{
Accessor for parent genotypes
}

