% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{parentLLPlot}
\alias{parentLLPlot}
\title{Create a diagnostic plot of a parentage inference, looking at all log likelihood ratios.}
\usage{
parentLLPlot(x, progeny)
}
\description{
Create a diagnostic plot of a parentage inference, looking at all log likelihood ratios.
}

