% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{parentNames,ProgenyArray-method}
\alias{parentNames,ProgenyArray-method}
\title{Return parent sample names}
\usage{
\S4method{parentNames}{ProgenyArray}(x)
}
\arguments{
\item{x}{a ProgenyArray object}
}
\description{
Return parent sample names
}

