% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/progenyarray-methods.R
\docType{methods}
\name{parentNames}
\alias{parentNames}
\title{Accessor for parents sample names}
\usage{
parentNames(x)

\S4method{parentNames}{ProgenyArray}(x) <- value
}
\arguments{
\item{x}{a ProgenyArray object}

\item{value}{vector of parent names}
}
\description{
Accessor for parents sample names

Set method for parents sample names
}

