% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{parents,ProgenyArray-method}
\alias{parents,ProgenyArray-method}
\title{Accessor for parents \code{data.frame} in a ProgenyArray object}
\usage{
\S4method{parents}{ProgenyArray}(x)
}
\arguments{
\item{x}{a ProgenyArray object}
}
\description{
Accessor for parents \code{data.frame} in a ProgenyArray object
}

