% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{ProgenyArray-class}
\alias{ProgenyArray-class}
\title{An S4 class that stores genotyping data from a progeny array design.}
\description{
An S4 class that stores genotyping data from a progeny array design.
}
\section{Slots}{

\describe{
\item{\code{ranges}}{a \code{GenomicRanges} object of loci}

\item{\code{ref}}{reference alleles}

\item{\code{alt}}{alternate alleles}

\item{\code{progeny}}{a matrix of bialleic progeny genotypes}

\item{\code{parents}}{a matrix of bialleic parents genotypes}

\item{\code{mothers}}{integer vector indicating the mother of each progeny}

\item{\code{fathers}}{integer vector indicating the father of each progeny}

\item{\code{samples}}{sample names}
}}

