% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/simulation-methods.R
\docType{methods}
\name{progenyArray,SimulatedProgenyArray-method}
\alias{progenyArray,SimulatedProgenyArray-method}
\title{Accessor for ProgenyArray object}
\usage{
\S4method{progenyArray}{SimulatedProgenyArray}(x)
}
\arguments{
\item{x}{a SimulationData object}
}
\description{
Accessor for ProgenyArray object
}

