% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{ProgenyArray}
\alias{ProgenyArray}
\alias{ProgenyArray-package}
\title{ProgenyArray.}
\usage{
ProgenyArray(progeny_geno, parents_geno, mothers = integer(),
  loci = GRanges(), ref = character(), alt = CharacterList())
}
\arguments{
\item{progeny_geno}{}

\item{parents_geno}{}

\item{loci}{a \code{GRanges} object of loci}

\item{ref}{a character vector of reference loci}

\item{alt}{a \code{CharacterList} of alternate loci}
}
\description{
ProgenyArray.

Create a new ProgenyArray object for storing genotyping and other data from
a progeny array experiment.
}

