% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{ProgenyArray}
\alias{ProgenyArray}
\alias{ProgenyArray-package}
\title{ProgenyArray.}
\usage{
ProgenyArray(progeny, parents, mothers = integer(), loci = GRanges(),
  ref = character(), alt = CharacterList(), samples = character())
}
\arguments{
\item{geno}{a genotype matrix}

\item{loci}{a \S4Class{GRanges} object of loci}

\item{ref}{a character vector of reference loci}

\item{alt}{a \S4Class{CharacterList} of alternate loci}

\item{a}{character vector of sample names}
}
\description{
ProgenyArray.

Create a new ProgenyArray object for storing genotyping and other data from
a progeny array experiment.
}

