% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{progenyGenotypes,ProgenyArray-method}
\alias{progenyGenotypes,ProgenyArray-method}
\title{Accessor for progeny genotypes}
\usage{
\S4method{progenyGenotypes}{ProgenyArray}(x)
}
\arguments{
\item{x}{a ProgenyArray object}
}
\description{
Accessor for progeny genotypes
}

