% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/progenyarray-methods.R
\docType{methods}
\name{progenyNames}
\alias{progenyNames}
\title{Accessor for progeny sample names}
\usage{
progenyNames(x)

\S4method{freqs}{ProgenyArray}(x) <- value

\S4method{progenyNames}{ProgenyArray}(x) <- value

\S4method{parentage}{ProgenyArray}(x)
}
\arguments{
\item{x}{a ProgenyArray object}

\item{value}{vector of parent names}

\item{x}{a ProgenyArray object}

\item{x}{a ProgenyArray object}
}
\description{
Accessor for progeny sample names

Manually set the allele frequencies

Set method for progeny sample names

Get inferred parents as a dataframe
}

