% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{ref,ProgenyArray-method}
\alias{ref,ProgenyArray-method}
\title{Accessor for reference alleles from a ProgenyArray object}
\usage{
\S4method{ref}{ProgenyArray}(x, as_char = TRUE)
}
\arguments{
\item{x}{a ProgenyArray object}

\item{as_char}{logical indicating whether to return as character
or as Tassel encodes them, as integer}
}
\description{
Accessor for reference alleles from a ProgenyArray object
}

