% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/progenyarray-methods.R
\docType{methods}
\name{show,ProgenyArray-method}
\alias{show,ProgenyArray-method}
\title{Pretty-print a ProgenyArray object}
\usage{
\S4method{show}{ProgenyArray}(object)
}
\arguments{
\item{object}{a ProgenyArray object}
}
\description{
Pretty-print a ProgenyArray object
}

