% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/simulation-methods.R
\name{SimulatedProgenyArray}
\alias{SimulatedProgenyArray}
\title{Create some simulation data}
\usage{
SimulatedProgenyArray(nparent, nprogeny, nloci, selfing = 0.5,
  prop_parent_missing = 0.2, prop_progeny_missing = 0.2, ehet = 0.8,
  ehom = 0.1)
}
\arguments{
\item{nparent}{number of parents}

\item{nprogeny}{number of progeny}

\item{nloci}{number of loci}

\item{selfing}{selfing rate}

\item{prop_parent_missing}{proportion of random missing data in parents}

\item{prop_progeny_missing}{proportion of random missing data in progeny}

\item{ehet}{heterozygous error rate}

\item{ehom}{homozygous error rate}
}
\description{
Create some simulation data
}

