% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{suppliedMothers,ProgenyArray-method}
\alias{suppliedMothers,ProgenyArray-method}
\title{Accessor for user-supplied mothers}
\usage{
\S4method{suppliedMothers}{ProgenyArray}(x)
}
\arguments{
\item{x}{a ProgenyArray object}
}
\description{
Accessor for user-supplied mothers
}

