% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/progenyarray-methods.R
\docType{methods}
\name{suppliedMothers}
\alias{suppliedMothers}
\title{Accessor for user-supplied mothers (returning indices or names)}
\usage{
suppliedMothers(x)

\S4method{suppliedMothers}{ProgenyArray}(x) <- value
}
\arguments{
\item{x}{a ProgenyArray object}

\item{value}{vector of mother indices}
}
\description{
Accessor for user-supplied mothers (returning indices or names)

Set method for supplied mothers
}

